+++ packages/create-deepspace/CHANGELOG.md
@@ -1,3 +1,5 @@
 # create-deepspace
 
+## 0.6.0
+
 ## 0.5.7